Add processor classes for KakaoMap, MyRealTrip, and TripDotCom
